# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Engineer Route Plan API
  version: 0.1.0
paths:
...
  /api/v1/plan:
    get:
      operationId: list_plans
      summary: Список планов региона
      description: >-
        Все планы выбранного региона (опционально — только одного набора
        бригад), в любом статусе running/done/failed, по убыванию created_at
        (сначала новые). Лёгкая сводка без маршрутов и метрик — детали
        конкретного плана отдаёт GET /plan/{plan_id}. Регион без загруженных
        данных — пустой список.
      parameters:
        - $ref: "#/components/parameters/RegionQuery"
        - $ref: "#/components/parameters/EngineerSetIdQuery"
      responses:
        "200":
          description: Планы региона (или одного набора бригад)
          headers:
            X-Request-ID:
              $ref: "./common.yaml#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanSummary"
        "400":
          $ref: "./common.yaml#/components/responses/BadRequest"
          description: >-
            region не из конфигурации или не передан, либо engineer_set_id не
            из этого региона (`fields`).
        "500":
          $ref: "./common.yaml#/components/responses/InternalError"
        "503":
          $ref: "./common.yaml#/components/responses/ServiceUnavailable"
          description: Недоступна база данных.
  /api/v1/plan/build:
    post:
      operationId: build_plan
      summary: Поставить построение плана выездов региона на дату в очередь
      description: >-
        Ставит построение плана в очередь и отвечает не дожидаясь его конца:
        назначение открытых (не completed и не cancelled) заявок региона на
        дату plan_date бригадам региона — основным алгоритмом (RoutingModel,
        трёхфазная лексикографическая оптимизация) или независимым baseline
        (FCFS, для сравнения) — и атрибуция причин отказа идут в фоне после
        ответа. Клиент получает plan_id сразу и узнаёт результат, опрашивая
        `GET /plan/{plan_id}`, пока status не станет done или failed. Каждый
        вызов — новый план (новый plan_id); построение не заменяет и не
        удаляет более ранние планы того же региона и даты.


        Регион должен иметь загруженные данные (`POST /data/upload` или
        `POST /data/demo`) — иначе 400. Каждая открытая заявка региона должна
        приходиться на дату plan_date (окно заявки — из файла одной загрузки,
        все заявки на один день) — несовпадение хотя бы одной заявки отклоняет
        весь запрос 400, план не ставится в очередь.


        `engineer_set_id` выбирает, по какому набору бригад региона строить
        план (без него — набор default); план хранит свой набор, и
        перепланирование этого плана идёт по бригадам того же набора.


        Число точек будущей маршрутной матрицы (бригады + заявки) проверяется
        до постановки в очередь — больше предела OSRM отклоняет запрос 400.
        Время в пути и расстояние считаются по дорожному графу (OSRM) уже в
        фоне, отдельно для каждого типа транспорта бригад региона. Основной
        алгоритм выполняется в одном процессе на весь сервер — конкурентное
        построение встаёт в очередь исполнителя, а не запускается параллельно
        вторым процессом.


        Плановое прибытие — расчётное время в пути от предыдущей точки
        маршрута (или от точки старта бригады) плюс текущее время; норматив
        «дорога 20 минут» здесь не используется. Причина каждой неназначенной
        заявки определяется процедурой атрибуции (навык → транспорт →
        окно/смена → «все подходящие бригады заняты»), а не первой
        проваленной проверкой.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanBuildRequest"
      responses:
        "202":
          description: План поставлен в очередь (status = running), маршрутов ещё нет
          headers:
            X-Request-ID:
              $ref: "./common.yaml#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "400":
          $ref: "./common.yaml#/components/responses/BadRequest"
          description: >-
            Регион не из конфигурации, у региона нет загруженных данных, окно
            хотя бы одной открытой заявки региона не на дату plan_date, либо
            число точек маршрутной матрицы больше предела OSRM.
        "413":
          $ref: "./common.yaml#/components/responses/PayloadTooLarge"
        "500":
          $ref: "./common.yaml#/components/responses/InternalError"
        "503":
          $ref: "./common.yaml#/components/responses/ServiceUnavailable"
          description: >-
            Недоступна база данных при постановке плана в очередь. План не
            создан.
  /api/v1/plan/{plan_id}:
    get:
      operationId: get_plan
      summary: Получить план — построенный, ещё считающийся или не сложившийся
      description: >-
        Возвращает план в любом из трёх состояний status без обращения к OSRM
        или солверу: running — план поставлен в очередь, маршрутов ещё нет;
        done — маршруты, объяснения и причины неназначенных заявок читаются
        из сохранённых данных построения; failed — построение не закончилось,
        причина — в failed_reason. Клиент опрашивает эту операцию, пока
        status не перестанет быть running.
      parameters:
        - name: plan_id
          in: path
          required: true
          description: Идентификатор плана на сервере
          schema:
            type: integer
            minimum: 1
            maximum: 9223372036854775807
      responses:
        "200":
          description: План в состоянии running, done или failed
          headers:
            X-Request-ID:
              $ref: "./common.yaml#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "400":
          $ref: "./common.yaml#/components/responses/BadRequest"
          description: plan_id вне допустимого диапазона.
        "404":
          $ref: "./common.yaml#/components/responses/NotFound"
          description: Плана с таким plan_id нет.
        "500":
          $ref: "./common.yaml#/components/responses/InternalError"
        "503":
          $ref: "./common.yaml#/components/responses/ServiceUnavailable"
          description: Недоступна база данных.
    delete:
      operationId: delete_plan
      summary: Удалить план вместе со всеми его перепланированиями
      description: >-
        Удаляет план и всю цепочку планов, порождённых от него
        `POST /plan/{plan_id}/replan` (прямо или через несколько
        перепланирований), вместе с их визитами и событиями перепланирования,
        в одной транзакции. План в состоянии running удалить нельзя — 409:
        его может ещё дописывать фоновое построение.
      parameters:
        - name: plan_id
          in: path
          required: true
          description: Идентификатор плана на сервере
          schema:
            type: integer
            minimum: 1
            maximum: 9223372036854775807
      responses:
        "204":
          description: План и вся цепочка его перепланирований удалены
          headers:
            X-Request-ID:
              $ref: "./common.yaml#/components/headers/X-Request-ID"
        "400":
          $ref: "./common.yaml#/components/responses/BadRequest"
          description: plan_id вне допустимого диапазона.
        "404":
          $ref: "./common.yaml#/components/responses/NotFound"
          description: Плана с таким plan_id нет.
        "409":
          $ref: "./common.yaml#/components/responses/Conflict"
          description: План ещё в состоянии running.
        "500":
          $ref: "./common.yaml#/components/responses/InternalError"
        "503":
          $ref: "./common.yaml#/components/responses/ServiceUnavailable"
          description: Недоступна база данных. План не удалён.
  /api/v1/plan/{plan_id}/compare:
    get:
      operationId: compare_plan
      summary: Сравнить обязательные метрики плана с baseline-планом
      description: >-
        Сравнивает план из пути запроса (обычно `algorithm=or_tools`) с
        baseline-планом (`baseline_plan_id`, обычно `algorithm=baseline_fcfs`)
        по обязательным метрикам (`engineers_used`, `total_distance_km`) —
        по одной записи на каждую. Оба плана должны быть в состоянии
        `status=done`: метрики строящегося или не сложившегося плана не
        определены. Оба плана должны быть одного набора бригад
        (`engineer_set_id`) — иначе сравнение бессмысленно: у планов разных
        наборов разное число бригад. `idle_time` в сравнение не входит — она
        только для отображения (см. `PlanMetrics.idle_time_by_engineer_min`).
      parameters:
        - name: plan_id
          in: path
          required: true
          description: Идентификатор плана на сервере
          schema:
            type: integer
            minimum: 1
            maximum: 9223372036854775807
        - name: baseline_plan_id
          in: query
          required: true
          description: Идентификатор baseline-плана, с которым сравнивается plan_id
          schema:
            type: integer
            minimum: 1
            maximum: 9223372036854775807
      responses:
        "200":
          description: Сравнение по каждой обязательной метрике, в порядке engineers_used, total_distance_km
          headers:
            X-Request-ID:
              $ref: "./common.yaml#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanComparisonEntry"
        "400":
          $ref: "./common.yaml#/components/responses/BadRequest"
          description: >-
            plan_id или baseline_plan_id вне допустимого диапазона (`fields`),
            либо один из двух планов ещё не done или у планов разные
            engineer_set_id (`message`).
        "404":
          $ref: "./common.yaml#/components/responses/NotFound"
          description: Плана с таким plan_id или baseline_plan_id нет.
        "500":
          $ref: "./common.yaml#/components/responses/InternalError"
        "503":
          $ref: "./common.yaml#/components/responses/ServiceUnavailable"
          description: Недоступна база данных.
  /api/v1/plan/{plan_id}/replan:
    post:
      operationId: replan_plan
      summary: Перепланировать бригады по одному событию дня
      description: >-
        Синхронно пересчитывает часть плана-родителя (`plan_id`) по одному
        событию перепланирования (Contract Net: объявление задания
        бригадам-кандидатам, локальные ставки, победитель с минимальной
        ставкой) и сохраняет результат новым планом с `parent_plan_id =
        plan_id`. Новый план наследует `engineer_set_id` плана-родителя —
        перепланирование пересчитывает бригады того же набора, набор сменить
        нельзя. План-родитель должен быть в состоянии `status=done` —
        иначе `400`. Пересчёт идёт от состояния бригад на момент
        `triggered_at` события: заявки `completed`/`cancelled` исключены,
        `in_progress` зафиксирована (бригада освобождается не раньше
        планового окончания этой заявки, начатая работа не прерывается),
        `en_route` и ещё не начатые заявки можно переставить. Затрагиваются
        только маршруты бригад, на которые повлияло событие — план остальных
        бригад переносится без изменений. Ответ несёт блок `diff` — что
        именно изменилось относительно `plan_id`.


        `new_urgent_ticket` — новая аварийная заявка. Кандидаты — бригады с
        навыком `emergency`; побеждает ставка с минимальным временем
        прибытия среди успевающих к `triggered_at + reaction_min` (60–120,
        по умолчанию 120), иначе — минимальное время прибытия среди всех
        кандидатов (превышение реакции отражается в `explanation` визита).
        Сначала — вставка в маршрут победителя без нарушения окон уже
        стоящих там заявок; если это невозможно — вытесняется одна не
        начатая заявка победителя более низкого приоритета, которая
        повторно торгуется среди остальных бригад без нарушения её
        собственного окна (каскад глубиной 1 — второе вытеснение не
        выполняется, вытесненная заявка, которую не удалось пристроить,
        уходит в `unassigned`). Если ни одна бригада не обладает навыком
        `emergency` — заявка уходит в `unassigned`. Заявке сервер сам
        присваивает `required_skill=emergency`, `priority=1`,
        `duration_min=80` (время на объекте), `received_at = triggered_at` и
        окно `[triggered_at, конец даты плана]` — окно здесь не сужает выбор
        бригады, целевое время реакции задаёт только `reaction_min`.


        `new_ticket` — новая обычная заявка в течение дня (необязательная
        возможность): только вставка в свободный интервал маршрута
        бригады-кандидата, без объявления задания и без вытеснения.
        `required_skill`, `priority` и `duration_min` определяются парой
        (`type_bk`, `type_hd`) по той же таблице соответствия типов, что и
        при загрузке файла — неизвестная пара типов отклоняется `400` ещё до
        поиска бригады. Кандидаты — бригады с этим навыком (и транспортом,
        если задан); среди их фактических маршрутов ищется свободный
        интервал, где прибытие на новую заявку попадает в её окно и не
        отодвигает ни один уже стоящий визит бригады; побеждает интервал с
        минимальным временем прибытия. Не найдено ни одного интервала —
        заявка уходит в `unassigned`, план остальных бригад не меняется.


        `ticket_cancelled` — заявка уже отменена (`PATCH
        /tickets/{ticket_id}/status`, `status=cancelled`) и убирается из
        маршрута бригады, на которую была назначена; последующие визиты этой
        бригады сдвигаются по времени; остальные бригады не пересчитываются.


        `engineer_unavailable` — бригада стала недоступна. С её маршрута
        снимаются все ещё не начатые визиты (`en_route` и позже не начатые;
        визит `in_progress` — как и у остальных событий — зафиксирован).
        Для каждой снятой заявки ищется свободный интервал в маршруте другой
        бригады региона — так же, как у `new_ticket`: без объявления задания
        и без вытеснения, побеждает интервал с минимальным временем прибытия.
        Не нашлось интервала ни у одной бригады — заявка уходит в
        `unassigned`. Сама ставшая недоступной бригада остаётся в ответе без
        визитов; последующие события снова могут назначать ей заявки.
      parameters:
        - name: plan_id
          in: path
          required: true
          description: Идентификатор плана-родителя на сервере
          schema:
            type: integer
            minimum: 1
            maximum: 9223372036854775807
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplanEventRequest"
      responses:
        "200":
          description: Новый план с diff относительно plan_id
          headers:
            X-Request-ID:
              $ref: "./common.yaml#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanReplanResult"
        "400":
          $ref: "./common.yaml#/components/responses/BadRequest"
          description: >-
            Событие не проходит ограничения контракта (`fields`), либо
            plan_id ещё не done, либо triggered_at раньше начала или позже
            конца даты плана, либо (`new_ticket`) window_start не раньше
            window_end, окно не приходится на дату плана, или пара
            (type_bk, type_hd) не найдена в таблице соответствия типов
            (`message`).
        "404":
          $ref: "./common.yaml#/components/responses/NotFound"
          description: >-
            Плана с таким plan_id нет, либо (для ticket_cancelled) заявки с
            таким ticket_id нет, либо (для engineer_unavailable) бригады с
            таким engineer_id нет.
        "409":
          $ref: "./common.yaml#/components/responses/Conflict"
          description: (ticket_cancelled) заявка ещё не в статусе cancelled.
        "500":
          $ref: "./common.yaml#/components/responses/InternalError"
        "503":
          $ref: "./common.yaml#/components/responses/ServiceUnavailable"
          description: Недоступны OSRM или база данных.
components:
  parameters:
    RegionQuery:
      name: region
      in: query
      required: true
      description: >-
        Код региона из списка регионов. Код, которого нет в конфигурации
        сервера, — ошибка 400 у параметра region.
      schema:
        $ref: "#/components/schemas/RegionCode"
    EngineerSetIdQuery:
      name: engineer_set_id
      in: query
      required: false
      description: >-
        Набор бригад региона. Без параметра — набор default. Набор, которого
        нет или который принадлежит другому региону, — ошибка 400 у параметра
        engineer_set_id.
      schema:
        type: integer
        minimum: 1
        maximum: 9223372036854775807
  schemas:
...
    RegionCode:
      type: string
      minLength: 1
      maxLength: 50
      pattern: "^[a-z][a-z0-9_]*$"
      description: Код региона — латиница в нижнем регистре, цифры и «_»; начинается с буквы
      examples: ["east"]
...
    PlanAlgorithm:
      type: string
      enum: [or_tools, baseline_fcfs]
      description: >-
        Чем строится план: or_tools — основной алгоритм (RoutingModel,
        трёхфазная лексикографическая оптимизация), baseline_fcfs — независимый
        план для сравнения (назначение по порядку поступления заявок)
    PlanBuildRequest:
      type: object
      additionalProperties: false
      required: [region, plan_date, algorithm]
      properties:
        region:
          $ref: "#/components/schemas/RegionCode"
        plan_date:
          $ref: "./common.yaml#/components/schemas/LocalDate"
          description: >-
            Дата плана; окно каждой открытой заявки региона должно приходиться
            на эту дату
        algorithm:
          $ref: "#/components/schemas/PlanAlgorithm"
        engineer_set_id:
          type: [integer, "null"]
          minimum: 1
          maximum: 9223372036854775807
          default: null
          description: >-
            Набор бригад региона, для которого строится план; null или
            отсутствует — набор default. Набор, которого нет или который
            принадлежит другому региону, — ошибка 400 у поля engineer_set_id
...
    PlanComparisonEntry:
      type: object
      additionalProperties: false
      required: [metric, main, baseline, delta]
      properties:
        metric:
          $ref: "#/components/schemas/PlanComparisonMetric"
        main:
          type: number
          description: Значение метрики у плана из пути запроса
        baseline:
          type: number
          description: Значение той же метрики у baseline-плана
        delta:
          type: number
          description: main минус baseline; отрицательное значение — план лучше baseline
    Plan:
      type: object
      additionalProperties: false
      required: [plan_id, algorithm, status, region, engineer_set_id]
      properties:
        plan_id:
          type: integer
          description: Идентификатор плана на сервере
        algorithm:
          $ref: "#/components/schemas/PlanAlgorithm"
        status:
          $ref: "#/components/schemas/PlanStatus"
        region:
          $ref: "#/components/schemas/RegionCode"
          description: >-
            Регион плана, каким бы ни был status — клиент, открывший план по
            прямой ссылке (GET /plan/{id} без предварительного выбора
            региона), берёт регион отсюда
        engineer_set_id:
          type: integer
          description: >-
            Набор бригад, для которого построен план — тот же, что был передан
            (или default) в POST /plan/build, каким бы ни был status
        engineers:
          anyOf:
            - type: array
              items:
                $ref: "#/components/schemas/EngineerRoute"
            - type: "null"
          description: >-
            Маршруты всех бригад региона на дату плана, по возрастанию engineer_id;
            null, если status не done
        unassigned:
          anyOf:
            - type: array
              items:
                $ref: "#/components/schemas/UnassignedTicket"
            - type: "null"
          description: >-
            Неназначенные заявки, по возрастанию ticket_id; null, если status не done
        metrics:
          anyOf:
            - $ref: "#/components/schemas/PlanMetrics"
            - type: "null"
          description: Обязательные метрики плана; null, если status не done
        failed_reason:
          anyOf:
            - $ref: "#/components/schemas/PlanFailedReason"
            - type: "null"
          description: Причина отказа; null, если status не failed
    PlanSummary:
      type: object
      additionalProperties: false
      required:
        - plan_id
        - region
        - engineer_set_id
        - plan_date
        - algorithm
        - status
        - created_at
        - parent_plan_id
        - failed_reason
      properties:
        plan_id:
          type: integer
          description: Идентификатор плана на сервере
        region:
          $ref: "#/components/schemas/RegionCode"
        engineer_set_id:
          type: integer
          description: Набор бригад, для которого построен план
        plan_date:
          $ref: "./common.yaml#/components/schemas/LocalDate"
          description: Дата плана
        algorithm:
          $ref: "#/components/schemas/PlanAlgorithm"
        status:
          $ref: "#/components/schemas/PlanStatus"
        created_at:
          $ref: "./common.yaml#/components/schemas/LocalDateTime"
          description: Когда построение плана поставлено в очередь
        parent_plan_id:
          anyOf:
            - type: integer
            - type: "null"
          description: >-
            План, от которого этот получен перепланированием; null — план
            построен POST /plan/build напрямую
        failed_reason:
          anyOf:
            - $ref: "#/components/schemas/PlanFailedReason"
            - type: "null"
          description: Причина отказа; null, если status не failed
...
    ReplanEventRequest:
      description: Одно событие перепланирования; тип определяет event_type.
      oneOf:
        - $ref: "#/components/schemas/NewUrgentTicketEvent"
        - $ref: "#/components/schemas/NewTicketEvent"
        - $ref: "#/components/schemas/TicketCancelledEvent"
        - $ref: "#/components/schemas/EngineerUnavailableEvent"
...